dont with customer airline and user

+++ back/models/Airlines.js
@@ -29,15 +29,21 @@ const AirLinesSchema = new mongoose.Schema({
 AirLinesSchema.plugin(AutoIncrement, {inc_field: 'airline_id'});
 
 
+
 // remove all the flight connected to the airline
 AirLinesSchema.pre('findOneAndRemove',async function(next) {
     
     // get the current airline details
     const current_airline = await this.model.findOne(this.getQuery())
-    const id = current_airline._id
-    await Flight.deleteMany({airline_company: id})
+
+    // in case there is no flight
+    if(current_airline){
+        const id = current_airline._id
+        await Flight.deleteMany({airline_company: id})
+    }
     next();
 });
+
 
 
 // // serializer for customer
